ai-agents/whatsapp-outreach-agent/requirements.txt
fastapi
uvicorn
python-dotenv
requests
beautifulsoup4
groq
twilio
